requirements.txt
requests==2.31.0
pyperclip==1.8.2
python-dotenv==1.0.0
openai==1.68.2
pydantic==2.10.6
Pillow
fastapi==0.110.0
uvicorn==0.27.1
